Rename hpp to h
I've forgotten the CRAN check god doesn't like hpp in `src/`, sorry.
