+++ src/adaR.h
